[nrf noup][tfm] Remove unused EdDSA calls to help reduce code size
Applied with:
git am 0004-Remove-unused-EdDSA-calls-to-help-reduce-code-size.patch

With the help of dead-code elimination and actual code removal
in this patch it's possible to skip the complete EdDSA algorithm
support from the build process.

Remove EdDSA from sign verifier as it does not need to be covered
by the tests. Remove it from the code so the build does not depend
on it.

Update known_good_test to prevent test failure due to omitted EdDSA
context initialization.